refactor(deployments): split env ui
